# src/main/java/com/clsa/controller/CsvController.java
package com.clsa.controller;

import javax.validation.Valid;

import org.apache.logging.log4j.LogManager;
import org.apache.logging.log4j.Logger;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;

import com.clsa.csvservice.api.ValidationApi;
import com.clsa.csvservice.models.CsvEntity;
import com.clsa.csvservice.models.ValidationResult;
import com.clsa.error.ApiErrorCode;
import com.clsa.service.CsvService;
import com.clsa.util.ApiUtil;

@RestController
/**
 * This is the main controller for CSV API
 * which implements the API interface generated by swagger codegen
 */
public class CsvController implements ValidationApi {
	
	private static final Logger logger = LogManager.getLogger(CsvController.class);
	
	private static final String SUCCESS_CODE = "SUCCESS-001";
	private static final String SUCCESS_DESC = "Validation is successful with no issues found";
	
	@Autowired
	private CsvService csvService;

	@Override
	public ResponseEntity<ValidationResult> csvValidation(@Valid CsvEntity body) {
		
		logger.info("entering csvValidation | body: {}", body);
		
		ValidationResult response = new ValidationResult();
		
		if (csvService.isCsvValid(body, response)) {
			response.setReturnCode(SUCCESS_CODE);
			response.setReturnDesc(SUCCESS_DESC);
		    return new ResponseEntity<>(response, HttpStatus.OK);
		} else {
			return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
		}	
	}
	
}